Move CommandBufferData to its own file
Moves CommandBufferData to its own .h and .cc files.
Removes "kind" from Cmd class and replaces it with
"IsDrawCall" function.
Some minor changes.
